feat: Add support for "prefers-color-scheme"
And fix some old styling bugs I hadn't noticed after migrating from
twind.

I used this as an opportunity to use `styles.css` for the primary colors,
as custom CSS properties, so that I can use them in the future if I have
to move from Tailwind. Another advantage to this is that I don't have to
write `dark:` or `light:` everywhere I have to define colors based on the
`prefers-color-scheme`.

Signed-off-by: Tim Hårek Andreassen <tim@harek.no>